backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
requests
beautifulsoup4
pymupdf
sqlalchemy
pytest
httpx
pillow
pytesseract
